Implement Audio Handling
This commit adds the scripts, assets, resources, and scene that I've
been using to implement and test audio handling.

As of right now, two octaves are available for playing, and Musical
Typing has been implemented as a debug function. Musical Typing
maps 19 keys (full octave + 7 extra notes) on the keyboard to
triggers for Note Events which allows for testing sound output
without having to have the actual peripheral or hard-coding events.

Musical Typing is only active if the preprocessor variables DEBUG &
DEBUG_MUSICAL_TYPING are set. If this push doesn't automatically
set these, then they can be set by going to
File->Build Settings->Player Settings and inserting
DEBUG;DEBUG_MUSICAL_TYPING into the Scripting Define Symbols field.

A scene that is used for testing audio is also included. It's
pretty empty right now, but it has a slider to set the velocity
that's used when the Musical Typing function fires off a Note
Event.

I'll create a new branch after this for further work on improving
the audio handling. I just figured that I might as well make an
initial commit.