Implement support for drum loops
This commit implements support for drum loops and making them in the
song creator. This is the last push to github before moving over to
Unity's cloud services. I would publish it there instead, but I don't
know exactly how to use it and it's too late in the evening to try
messing with it.